Deep optimise: 172 fixes across 7 runs — security, i18n, bugs, naming, logging, dead code
Runs 1-7 of /optimise audited the full codebase (Kotlin, Express API, admin panel)
until a clean run found 0 remaining issues. Key changes:

- Security (20): CORS allowlist, mass assignment whitelists, XSS escaping,
  path traversal guards, rate limiting, generic error responses
- i18n (57): Hardcoded strings externalized to stringResource() with 19 locales
- Bugs (19): Integer overflow guards, null safety, auth middleware hardening
- Naming (31): Single-letter vars renamed across ViewModels, cron jobs, utils
- Logging (14): Silent catches replaced with warn logging, structured logging
- Dead code (7): Removed unused functions, fields, and stale migration code
- Bandwidth (5): Firestore .select() projections, Cache-Control headers
- Web (8): Accessibility labels, interval cleanup, escapeHtml

Tests: 331 passing (3 pre-existing failures unchanged), 12 new tests added.

+++ shared/src/commonMain/kotlin/com/shyden/shytalk/feature/gifting/GiftingViewModel.kt
@@ -139,7 +139,7 @@ class GiftingViewModel(
 
         if (!isBackpackTab) {
             val gift = state.giftCatalog.find { it.id == giftId } ?: return
-            val totalCost = gift.coinValue.toLong() * quantity * recipients.size
+            val totalCost = gift.coinValue.toLong() * quantity.toLong() * recipients.size
             if (totalCost > state.coinBalance) {
                 _uiState.update { it.copy(showConfirmDialog = false, navigateToWallet = true) }
                 return
